more types and casts
git-svn-id: https://svn.r-project.org/R/trunk@59171 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/deparse.c
@@ -1182,7 +1182,7 @@ static void print2buff(const char *strng, LocalParseData *d)
     bufflen = strlen(d->buffer.data);
     R_AllocStringBuffer(bufflen + tlen, &(d->buffer));
     strcat(d->buffer.data, strng);
-    d->len += tlen;
+    d->len += (int) tlen; // FIXME
 }
 
 static void vector2buff(SEXP vector, LocalParseData *d)
